move dace_lazy_stencil (review comment)
