+++ src/gt4py/gtscript.py
@@ -31,7 +31,7 @@ from gt4py.lazy_stencil import LazyStencil
 
 
 try:
-    from gt4py.dace_lazy_stencil import DaCeLazyStencil
+    from gt4py.backend.dace_lazy_stencil import DaCeLazyStencil
 except ImportError:
     DaCeLazyStencil = LazyStencil
